backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
passlib[argon2]>=1.7.4
argon2-cffi>=21.3.0
python-multipart>=0.0.6
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
starlette>=0.27.0
anyio>=3.7.0
